LFOPTEE-16: fix warning where uint32_t is expected
Fix warning where TEE_AEEncryptFinal() expects a uint32_t* instead of a
size_t*.

+++ ta/crypto_perf/ta_alg_ae.c
@@ -18,7 +18,7 @@ static TEE_OperationHandle authenDec_op;
 
 static uint8_t tag[64];
 static size_t  tag_len;
-static size_t  gen_tag_len;
+static uint32_t gen_tag_len;
 static uint8_t nonce[15];
 static size_t  nonce_len;
 static uint8_t aad[16];
